get_mentions/requirements.txt
python-decouple==3.6
redis==4.4.0
tweepy==4.12.1